refactor: Split up tests into two seperate files
