+++ generatorTests/test/fileStructureSpec.js
@@ -6,22 +6,7 @@ const ROOT_DIR = path.join(process.cwd(), '..');
 
 describe('As a dev', function() {
 
-    describe('When using the static site generator', function() {
-
-        it('then _config folder should exist', function() {
-            var pathToTest = path.join(ROOT_DIR, '_config');
-            fs.statSync(pathToTest);
-        })
-
-        it('then routes folder should exist', function() {
-            var pathToTest = path.join(ROOT_DIR, 'routes');
-            fs.statSync(pathToTest);
-        })
-
-        it('then views folder should exist', function() {
-            var pathToTest = path.join(ROOT_DIR, 'views');
-            fs.statSync(pathToTest);
-        })
+    describe('When testing for generator file structure', function() {
 
         it('then _config files should exist', function() {
             var pathToTest = path.join(ROOT_DIR, '_config/');
@@ -62,6 +47,7 @@ describe('As a dev', function() {
                 fs.statSync(path.join(ROOT_DIR, fileName));
             });
         })
+
     });
 
 });
